Extract CollectionCell into own file
